requirements.txt
streamlit
pyyaml
litellm